check the conserved motifs in the orthologs

Print the conserved motifs in the alignments

conservedMotifs(
  aln,
  aln_list,
  PWMs,
  queryGenome,
  background = "genome",
  ...,
  output_folder,
  format = c("txt", "html")
)

Arguments

aln

alignment of multiple DNAs. Output of alignment function.

aln_list

The list of output of searchTFBPS such as for human and mouse.

PWMs

The Position Weight Matrix list represented as a numeric matrix. Object of PWMatrixList or PFMatrixList.

queryGenome

An object of BSgenome for query enhancer.

background

Background nucleotide frequencies. Default is "genome". Refer matchMotifs for details.

...

Other parameters can be passed to to matchMotifs.

output_folder

Output folder name.

format

The format of output files with motif match positions. Available formats are 'txt' and 'html'. Default is 'txt'.

Value

A list of XStringViews

Examples

library(BSgenome.Hsapiens.UCSC.hg38)
library(BSgenome.Mmusculus.UCSC.mm10)
library(BSgenome.Drerio.UCSC.danRer10)
LEN <- GRanges("chr4", IRanges(19050041, 19051709))
seqEN <- getSeq(BSgenome.Drerio.UCSC.danRer10, LEN)
aln_hs <- readRDS(system.file("extdata", "aln_hs.rds",
               package="enhancerHomologSearch"))
genome(aln_hs) <- Hsapiens
aln_mm <- readRDS(system.file("extdata", "aln_mm.rds",
               package="enhancerHomologSearch"))
genome(aln_mm) <- Mmusculus
al <- alignment(seqEN, list(human=aln_hs, mouse=aln_mm),
                method="ClustalW", order="input")
data(motifs)
conservedMotifs(al[[1]], list(human=aln_hs, mouse=aln_mm),
                motifs[["dist60"]], Drerio)
#> $Enhancer
#> Views on a 1669-letter DNAString subject
#> subject: TGGCATACACAGCAAACATCATGAATTTAATTTA...TTAGATAAATAGAAACAGAAGCAAATTGGCGAGT
#> views:
#>       start  end width
#>   [1]  1563 1577    15 [AAAGATCAAATAGTT]
#>   [2]  1616 1635    20 [AAGTCTTAAATAAATTGTGA]
#>   [3]   469  481    13 [TGGTGGAATGTAA]
#>   [4]   399  406     8 [TTCACGTG]
#> 
#> $`human_chr7:128264261-128265260:+`
#> Views on a 1000-letter DNAString subject
#> subject: GAACAGAGAGGATTGCCTGGAGGTTCCTAGGACC...TTACGCTTTTTGTAAAGCAATCAGAACAGCCATC
#> views:
#>       start end width
#>   [1]   886 900    15 [AGAGATCAAATGTAA]
#>   [2]   893 912    20 [AAATGTAATAACGTGTGTGA]
#>   [3]   275 287    13 [GGAAGGAATGTCC]
#>   [4]   187 194     8 [CACGTGAC]
#> 
#> $`mouse_chr6:29032045-29033044:+`
#> Views on a 1000-letter DNAString subject
#> subject: TCTGAACTGCCTTGTTTACCAGCTGTCTTGCTAA...CGATACTTCTTTACCAACATCGGTGGTGGCATCT
#> views:
#>       start end width
#>   [1]    80  94    15 [AGAGCTTGGATCTTT]
#>   [2]   816 835    20 [CCACAGCATTTTTAACATTT]
#>   [3]   628 640    13 [ATCAGGAATGTGG]
#>   [4]   179 186     8 [CTCGTGAC]
#>