Transcription Factor Binding Pattern Similarity (TFBPS) search

Search the TFBPs for query in subject.

searchTFBPS(
  query,
  subject,
  PWMs,
  queryGenome,
  background = "genome",
  ...,
  maximalShuffleEnhancers = 1000
)

Arguments

query

An object of DNAStringSet to represent enhancer

subject

Output of getENCODEdata. An object of Enhancers

PWMs

The Position Weight Matrix list represented as a numeric matrix. Object of PWMatrixList or PFMatrixList.

queryGenome

An object of BSgenome for query data.

background

background nucleotide frequencies. Default is "genome". Refer matchMotifs for details.

...

Parameters will be passed to matchMotifs except 'out' and 'genome'.

maximalShuffleEnhancers

The maximal number of Shuffled enhancers. If the number of the input enhancer candidates is greater than maximalShuffleEnhancers, no shuffled enhancer sequences will be included. The shuffled enhancers will be created by shuffle.

Value

An object of Enhancers.

Examples

library(BSgenome.Hsapiens.UCSC.hg38)
peaks <- GRanges("chr1", IRanges(seq(5000, 50000, by=1000), width=1000))
peaks$id <- paste(seq_along(peaks), 1, sep="_")
subj <- Enhancers(genome=Hsapiens, peaks=peaks)
q <- getSeq(Hsapiens, GRanges("chr1", IRanges(90000, width=1000)))
data(motifs)
ao <- searchTFBPS(q, subj, motifs[["dist60"]], queryGenome=Hsapiens,
                  maximalShuffleEnhancers = 50)